requirements.txt
# Core Agent Framework (simplified for Python 3.8)
# Note: Using compatible versions that work with Python 3.8+

# LLM Integration
anthropic>=0.20.0
crewai>=0.100.0
langchain-openai>=0.1.6
langchain>=0.1.13

# Data Processing & Analysis (compatible versions)
numpy>=1.24.0
pandas>=2.0.0
python-dateutil>=2.8.0

# Machine Learning & Forecasting
scikit-learn>=1.3.0
prophet>=1.1
statsmodels>=0.14.0

# AWS Integration
boto3>=1.26.0

# Data Validation
pydantic>=2.0.0

# Environment Management
python-dotenv>=1.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# API Framework
fastapi>=0.95.0
uvicorn>=0.20.0

# Utilities
requests>=2.31.0
click>=8.0.0
